WIP: Moving times & hands component from face to analog clock components
